refactor: organize ChatBox utility modules
