+++ projects/app/src/components/core/chat/ChatContainer/ChatBox/utils/requestVariables.ts
