+++ projects/app/src/components/core/chat/ChatContainer/ChatBox/utils/scrollUtils.ts
